correcion de producto y agregado de maximo y minimo de stock ademas se agrego rfc al cliente

+++ app/Http/Controllers/Api/V1/ClienteController.php
@@ -104,6 +104,9 @@ class ClienteController extends Controller
         if(!$data) {
             return response()->json(['message'=> 'Registro no encontrado'],404);
         }
+        if ($data->ventas()->exists()) {
+            return response()->json(['message' => 'No se puede eliminar. El cliente tiene ventas registradas.'], 400);
+        }
         $data->delete();
         return response()->json(['message'=> 'Registro eliminado'],200);
     }
